Implement Mavlink custom_mode; Set mav system type based on # of motors configured; Move servo_output msg to rc_channels stream.
git-svn-id: http://autoquad.googlecode.com/svn/trunk@178 69e9ae41-46e5-6d7e-5b29-2a8ba9409dbb

+++ aq_mavlink.h
@@ -50,7 +50,7 @@ typedef struct {
     uint16_t packetDrops;
     uint16_t idlePercent;
     uint8_t mode;
-    uint8_t nav_mode;
+    uint32_t nav_mode;
     uint8_t status;
     uint8_t wpTargetSysId;
     uint8_t wpTargetCompId;
